Refactor: SkillReader 클래스 패키지 경로 변경
