+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/skill/core/SkillReader.java
@@ -1,8 +1,6 @@
-package kernel.jdon.moduleapi.domain.skill.application;
+package kernel.jdon.moduleapi.domain.skill.core;
 
 import java.util.List;
-
-import kernel.jdon.moduleapi.domain.skill.core.SkillInfo;
 
 public interface SkillReader {
 	List<SkillInfo.FindHotSkill> findHotSkillList();
