+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/skill/core/SkillServiceImpl.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kernel.jdon.moduleapi.domain.skill.application.SkillReader;
 import lombok.RequiredArgsConstructor;
 
 @Service
